Initial project structure
This commit introduces an initial README file that tries to give a good description of the project in it's current state. It is important for this file to be as up to date as possible.

This commit also introduces the core module which will contain low level functionalities that should be shared across the engine and which are difficult to categorize on their own.