Add GPU resource abstractions
Ng uses OpenGL 4.5 core for it's rendering, but I want to hide it as implementation details. So to help manage these resources lifetimes and hide this implementation details I added multiple RAII classes on top of various OpenGL resources.

The class gpu_context is a factory to create the resources so I can make sure no GPU resources are made before the OpenGL context is ready.