Add application class to handle the lifecycle of the application.
The host dispatch events from the user into a platform agnostic format to the application class.
At the moment, mouse events and keyboard events are supported. Other types of events will be added at a later time when needed.

It's also important to note that the application ticks at a fixed time step of 60 FPS but renders as fast as possible.