Added a name class that represent a more efficient string when comparison is needed.
The name will hash the content of its string so accelerate comparison. Before comparing the content of two strings to check if both are the same, the name will compare the hashes of the two strings and when both are equals, the content of the strings will be compared.

The strings are compared to be sure hash collisions doesn't have any impact in the comparison. It is possible to get rid of the string comparison if hash collision has no repercussion.